Add script to downscale 4k/8k textures for reduced VRAM usage in SkyrimVR
Lots of Skyrim mods add unnecessarily detailed textures, that just eat
up a bunch of VRAM and cause micro-stutters while looking around as the
driver has to constantly page textures in and out of the GPU once the
VRAM limit is reached.

This is a dead-simple cygwin script intended to be run from the
Data/textures directory in Skyrim that will locate any 4K / 8K loose
textures and run Microsoft texconv to downscale them to a much more
reasonable 2K.

The downsized textures are placed in a 2K directory that is intended to
then be turned into a mod and added to your mod organiser of choice to
allow it to handle overriding the 4K textures rather than just outright
overwriting the mod added textures directly, though the downside is that
switching mods to others using different textures would then be using
the wrong ones... This is for advanced users only who know what they are
doing.

Microsoft texcov can be downloaded from here and by default the script
expects it to be placed/linked in the cygwin home directory:

    https://github.com/microsoft/DirectXTex/releases