bashrc: Add cygwin implementations of pgrep/pkill
I use pkill all the time, but want it to affect Windows apps as well
when running under cygwin. This implements a bare bones wrapper around
Windows tasklist/taskkill to achieve this. A Windows pgrep is also
provided, though of lesser use since the PIDs given will be Windows PIDs
that cannot be used with the cygwin kill command.