bashrc: Don't use git prompt on Mac with broken git
If a Mac user uninstalls Xcode, it leaves a git binary in place that
will pop a dialog asking to install the command line tools whenever it
is run, which is very annoying if the git prompt is trying to run it
(and further aggravated by the fact that choosing to install the command
line tools from this dialog fails spectacularly).

If on Mac, make sure that Xcode is installed before attempting to enable
the git prompt.