bashrc/cygwin: Add a sudo implementation for cygwin
This will download and run Chronial's sudo implementation for cygwin,
which is generally superior to using cygwin's cygstart --action=runas
functionality since (barring the UAC prompt) it runs in the current
terminal, and properly parses the command line, etc. and is far more
convenient than opening a separate cygwin session as admin.