bash: Add p4c command to auto switch perforce workspace
The perforce command line client lacks a sensibly easy means to change
to the workspace appropriate for the current working directory, which is
a royal PITA when switching projects frequently.

This adds a p4c function into bash to determine which workspace root is
[above] the current working directory and matches the current perforce
user and host, then switches to it.