tmux: Use a cyan status line on remote sessions
Intention is to be able to quickly tell at a glance whether I'm typing
into the local machine or a remote one. This will update the status line
colour based on whether the SSH_CONNECTION environment variable is set
or not when attaching a client. It isn't perfect as I can still move
from one machine to another without [de]attaching clients, but most of
the time I have the remote machine detached so it's probably good enough
for now.