mouse alt+tab: Add a context menu item to elevate privileges
It's necessary to elevate the script privileges to allow it to send
alt+tab when a privileged program is in the foreground. We don't want to
do this by default to avoid a UAC every time the script is run, but add
a context menu to easily escalate if required.